src/app/file-upload/file-upload.component.ts
import { Component } from '@angular/core';

@Component({
  selector: 'app-file-upload',
  templateUrl: './file-upload.component.html',
  styleUrls: ['./file-upload.component.css'],
})
export class FileUploadComponent {
  public data: string;

  public onChange(event): void {
    const file = event.dataTransfer ? event.dataTransfer.files[0] : event.target.files[0];
    const reader = new FileReader();
    reader.onload = this.handleReaderLoaded.bind(this);
    reader.readAsDataURL(file);
  }

  private handleReaderLoaded(event): void {
    let file = event.target.result;
    file = file.split(',')[1];
    this.data = 'data: ' + window.atob(file);
  }
}
